+++ lib/Server.js
@@ -1,5 +1,7 @@
 'use strict';
 
+const path = require('path');
+const fs = require('fs');
 const url = require('url');
 const ipaddr = require('ipaddr.js');
 const internalIp = require('internal-ip');
@@ -72,6 +74,79 @@ class Server {
     this.webSocketProxies.forEach(function (webSocketProxy) {
       this.server.on('upgrade', webSocketProxy.upgrade);
     }, this);
+  }
+
+  static get DEFAULT_STATS() {
+    return {
+      all: false,
+      hash: true,
+      assets: true,
+      warnings: true,
+      errors: true,
+      errorDetails: false,
+    };
+  }
+
+  static getFreePort(port) {
+    const pRetry = require('p-retry');
+    const portfinder = require('portfinder');
+
+    if (port && port !== 'auto') {
+      return Promise.resolve(port);
+    }
+
+    function runPortFinder() {
+      return new Promise((resolve, reject) => {
+        // default port
+        portfinder.basePort = 8080;
+        portfinder.getPort((error, foundPort) => {
+          if (error) {
+            return reject(error);
+          }
+
+          return resolve(foundPort);
+        });
+      });
+    }
+
+    // Try to find unused port and listen on it for 3 times,
+    // if port is not specified in options.
+    const defaultPortRetry = parseInt(process.env.DEFAULT_PORT_RETRY, 10) || 3;
+
+    return pRetry(runPortFinder, { retries: defaultPortRetry });
+  }
+
+  static findCacheDir() {
+    const cwd = process.cwd();
+
+    let dir = cwd;
+
+    for (;;) {
+      try {
+        if (fs.statSync(path.join(dir, 'package.json')).isFile()) break;
+        // eslint-disable-next-line no-empty
+      } catch (e) {}
+
+      const parent = path.dirname(dir);
+
+      if (dir === parent) {
+        // eslint-disable-next-line no-undefined
+        dir = undefined;
+        break;
+      }
+
+      dir = parent;
+    }
+
+    if (!dir) {
+      return path.resolve(cwd, '.cache/webpack-dev-server');
+    } else if (process.versions.pnp === '1') {
+      return path.resolve(dir, '.pnp/.cache/webpack-dev-server');
+    } else if (process.versions.pnp === '3') {
+      return path.resolve(dir, '.yarn/.cache/webpack-dev-server');
+    }
+
+    return path.resolve(dir, 'node_modules/.cache/webpack-dev-server');
   }
 
   applyDevServerPlugin() {
@@ -539,7 +614,7 @@ class Server {
     });
   }
 
-  openBrowser(uri, options) {
+  openBrowser(uri) {
     const isAbsoluteUrl = require('is-absolute-url');
     const open = require('open');
 
@@ -579,12 +654,12 @@ class Server {
       });
     };
 
-    if (Array.isArray(options)) {
-      options.forEach((item) => {
+    if (Array.isArray(this.options.open)) {
+      this.options.open.forEach((item) => {
         openTasks.push(...getOpenTask(item));
       });
     } else {
-      openTasks.push(...getOpenTask(options));
+      openTasks.push(...getOpenTask(this.options.open));
     }
 
     Promise.all(
@@ -621,6 +696,25 @@ class Server {
         });
       })
     );
+  }
+
+  runBonjour() {
+    const bonjour = require('bonjour')();
+    const os = require('os');
+
+    bonjour.publish({
+      name: `Webpack Dev Server ${os.hostname()}:${this.options.port}`,
+      port: this.options.port,
+      type: this.options.https ? 'https' : 'http',
+      subtypes: ['webpack'],
+      ...this.options.bonjour,
+    });
+
+    process.on('exit', () => {
+      bonjour.unpublishAll(() => {
+        bonjour.destroy();
+      });
+    });
   }
 
   logStatus() {
@@ -747,7 +841,7 @@ class Server {
     if (this.options.open) {
       const openTarget = prettyPrintUrl(this.options.host || 'localhost');
 
-      this.openBrowser(openTarget, this.options.open);
+      this.openBrowser(openTarget);
     }
   }
 
@@ -806,23 +900,7 @@ class Server {
             }
 
             if (this.options.bonjour) {
-              const bonjour = require('bonjour')();
-              const os = require('os');
-              const { https, port } = this.options;
-
-              bonjour.publish({
-                name: `Webpack Dev Server ${os.hostname()}:${port}`,
-                port,
-                type: https ? 'https' : 'http',
-                subtypes: ['webpack'],
-                ...this.options.bonjour,
-              });
-
-              process.on('exit', () => {
-                bonjour.unpublishAll(() => {
-                  bonjour.destroy();
-                });
-              });
+              this.runBonjour();
             }
 
             this.logStatus();
@@ -861,47 +939,6 @@ class Server {
         this.middleware.close(callback);
       });
     });
-  }
-
-  static get DEFAULT_STATS() {
-    return {
-      all: false,
-      hash: true,
-      assets: true,
-      warnings: true,
-      errors: true,
-      errorDetails: false,
-    };
-  }
-
-  static getFreePort(port) {
-    const pRetry = require('p-retry');
-    const portfinder = require('portfinder');
-
-    if (port && port !== 'auto') {
-      return Promise.resolve(port);
-    }
-
-    function runPortFinder() {
-      return new Promise((resolve, reject) => {
-        // default port
-        portfinder.basePort = process.env.WEBPACK_DEV_SERVER_BASE_PORT || 8080;
-        portfinder.getPort((error, foundPort) => {
-          if (error) {
-            return reject(error);
-          }
-
-          return resolve(foundPort);
-        });
-      });
-    }
-
-    // Try to find unused port and listen on it for 3 times,
-    // if port is not specified in options.
-    const defaultPortRetry =
-      parseInt(process.env.WEBPACK_DEV_SERVER_PORT_RETRY, 10) || 3;
-
-    return pRetry(runPortFinder, { retries: defaultPortRetry });
   }
 
   getStats(statsObj) {
