--- lib/utils/findCacheDir.js
@@ -1,32 +1,0 @@
-'use strict';
-
-const fs = require('fs');
-const path = require('path');
-
-const findCacheDir = () => {
-  const cwd = process.cwd();
-  let dir = cwd;
-  for (;;) {
-    try {
-      if (fs.statSync(path.join(dir, 'package.json')).isFile()) break;
-      // eslint-disable-next-line no-empty
-    } catch (e) {}
-    const parent = path.dirname(dir);
-    if (dir === parent) {
-      // eslint-disable-next-line no-undefined
-      dir = undefined;
-      break;
-    }
-    dir = parent;
-  }
-  if (!dir) {
-    return path.resolve(cwd, '.cache/webpack-dev-server');
-  } else if (process.versions.pnp === '1') {
-    return path.resolve(dir, '.pnp/.cache/webpack-dev-server');
-  } else if (process.versions.pnp === '3') {
-    return path.resolve(dir, '.yarn/.cache/webpack-dev-server');
-  }
-  return path.resolve(dir, 'node_modules/.cache/webpack-dev-server');
-};
-
-module.exports = findCacheDir;
