+++ lib/utils/getCertificate.js
@@ -4,14 +4,14 @@ const path = require('path');
 const os = require('os');
 const fs = require('graceful-fs');
 const del = require('del');
-const findCacheDir = require('./findCacheDir');
+const Server = require('../Server');
 const createCertificate = require('./createCertificate');
 
 function getCertificate(logger) {
   // Use a self-signed certificate if no certificate was configured.
   // Cycle certs every 24 hours
   const certificateDir =
-    findCacheDir({ name: 'webpack-dev-server' }) || os.tmpdir();
+    Server.findCacheDir({ name: 'webpack-dev-server' }) || os.tmpdir();
   const certificatePath = path.join(certificateDir, 'server.pem');
 
   let certificateExists = fs.existsSync(certificatePath);
